Split out connectSocialUser
